refactor: create UserDefaults.swift
