+++ IVPNWidget/Scene/MainView.swift
@@ -4,7 +4,7 @@
 //  https://github.com/ivpn/ios-app
 //
 //  Created by Juraj Hilje on 2023-04-05.
-//  Copyright (c) 2021 Privatus Limited.
+//  Copyright (c) 2023 Privatus Limited.
 //
 //  This file is part of the IVPN iOS app.
 //
